chore: Add file_obj attribute to MMCIFParser
This commit adds a new `file_obj` attribute to the `MMCIFParser` class. This attribute is used to store the file object passed to the `parse` method. The `file_obj` is then used in the parsing process to read lines from the file. This change ensures that the `MMCIFParser` class has access to the file object throughout the parsing process, allowing for more flexibility in handling different types of file inputs.